perf test: test_intel_pt.sh: Fix return checking
Bugzilla: https://bugzilla.redhat.com/2177180

upstream
========
commit fd9b45e39cfaf885a8767bcb7631868155a2f4d6
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Mon Sep 12 11:34:09 2022 +0300

description
===========
The use of set -e will cause a function that returns non-zero to terminate
the script unless the result is consumed by || for example. That is OK if
there is only 1 test function, but not if there are more. Prepare for more
by using ||.

    Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
    Acked-by: Namhyung Kim <namhyung@kernel.org>
    Cc: Ian Rogers <irogers@google.com>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Link: https://lore.kernel.org/r/20220912083412.7058-9-adrian.hunter@intel.com
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>