redhat/configs: Enable WERROR only for gcc 11
Bugzilla: https://bugzilla.redhat.com/2184097
Upstream Status: RHEL only

RHEL9 is built with gcc 11 compiler and previous patches fixes
warnings that are specific to that compiler verison. New gcc compilers
may have new warnings which can fail the compilation process with
CONFIG_WERROR. To avoid this problem, we restrict the the use of
CONFIG_WERROR only if gcc 11 is used.

Suggested-by: Stefan Assmann <sassmann@redhat.com>
Signed-off-by: Waiman Long <longman@redhat.com>