perf test: Add arm64 asm pureloop test shell script
Bugzilla: https://bugzilla.redhat.com/2177180

upstream
========
commit fdc25cc59c7126952f04beafb0de6143a7fa574d
Author: Carsten Haitzler <carsten.haitzler@arm.com>
Date: Fri Sep 9 16:27:54 2022 +0100

description
===========
Add a script to drive the asm pureloop test for arm64/CoreSight that
gathers data so it passes a minimum bar for amount and quality of
content that we extract from the kernel's perf support.

Committer notes:

Add the install of tests/shell/coresight/*.sh to
tools/perf/Makefile.perf as we're starting to populate that dir.

    Reviewed-by: James Clark <james.clark@arm.com>
    Signed-off-by: Carsten Haitzler <carsten.haitzler@arm.com>
    Cc: Leo Yan <leo.yan@linaro.org>
    Cc: Mathieu Poirier <mathieu.poirier@linaro.org>
    Cc: Mike Leach <mike.leach@linaro.org>
    Cc: Suzuki Poulouse <suzuki.poulose@arm.com>
    Cc: coresight@lists.linaro.org
    Link: https://lore.kernel.org/r/20220909152803.2317006-5-carsten.haitzler@foss.arm.com
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>