perf bpf: Fix build with libbpf 0.7.0 by adding prototype for bpf_load_program()
Bugzilla: https://bugzilla.redhat.com/2177180

upstream
========
commit 409fb6bdd6eaf64c2da174b48b30ae032a9d7554
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Oct 18 21:47:51 2022 -0300

description
===========
The bpf_load_program() prototype appeared in tools/lib/bpf/bpf.h as
deprecated, but nowadays its completely removed, so add it back for
building with the system libbpf when using 'make LIBBPF_DYNAMIC=1'.

This is a stop gap hack till we do like tools/bpf does with bpftool,
i.e. bootstrap the libbpf build and install it in the perf build
directory when not using 'make LIBBPF_DYNAMIC=1'.

That has to be done to all libraries in tools/lib/, so tha we can
remove -Itools/lib/ from the tools/perf CFLAGS.

Noticed when building with LIBBPF_DYNAMIC=1 and libbpf 0.7.0 on RHEL9.

    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>