perf test: Use a test program in 'perf record' tests
Bugzilla: https://bugzilla.redhat.com/2177180

upstream
========
commit 4321ad4ee98b7325d6133e1d5b7fa25bcbdeb57e
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Oct 20 10:26:38 2022 -0700

description
===========
If the system has cc it could build a test program with two threads
and then use it for more detailed testing.  Also it accepts an option
to run a thread forever to ensure multi-thread runs.

If cc is not found, it falls back to use the default value 'true'.

    Reviewed-by: Adrian Hunter <adrian.hunter@intel.com>
    Signed-off-by: Namhyung Kim <namhyung@kernel.org>
    Acked-by: Ian Rogers <irogers@google.com>
    Cc: Ingo Molnar <mingo@kernel.org>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Link: https://lore.kernel.org/r/20221020172643.3458767-4-namhyung@kernel.org
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>