perf test: Test record with --threads option
Bugzilla: https://bugzilla.redhat.com/2177180

upstream
========
commit 7f4ed3f0b1fc951bcb91a78bf68ac45aac13ec0b
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Oct 20 10:26:42 2022 -0700

description
===========
The --threads option changed the 'perf record' behavior significantly,
so it'd be nice if we test it separately.  Add --threads options with
different argument in each test supported and check the result.

Also update the cleanup routine because threads recording produces data
in a directory.

    Reviewed-by: Adrian Hunter <adrian.hunter@intel.com>
    Signed-off-by: Namhyung Kim <namhyung@kernel.org>
    Acked-by: Ian Rogers <irogers@google.com>
    Cc: Ingo Molnar <mingo@kernel.org>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Link: https://lore.kernel.org/r/20221020172643.3458767-8-namhyung@kernel.org
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>