perf bpf: Rename perf_include_dir to libbpf_include_dir
Bugzilla: https://bugzilla.redhat.com/2177180

upstream
========
commit b018899e620b8ee4529f43bd02e9e8e43043e33e
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Nov 4 11:45:36 2022 -0300

description
===========
As this is where we expect to find bpf/bpf_helpers.h, etc.

This needs more work to make it follow LIBBPF_DYNAMIC=1 usage, i.e. when
not using the system libbpf it should use the headers in the in-kernel
sources libbpf in tools/lib/bpf.

We need to do that anyway to avoid this mixup system libbpf and
in-kernel files, so we'll get this sorted out that way.

And this also may become moot as we move to using BPF skels for this
feature.

    Cc: Adrian Hunter <adrian.hunter@intel.com>
    Cc: Ian Rogers <irogers@google.com>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Cc: Namhyung Kim <namhyung@kernel.org>
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>