perf lock contention: Do not use BPF task local storage
Bugzilla: https://bugzilla.redhat.com/2177180

upstream
========
commit c66a36af7ba3a628453da1d91f42ee64fb36ea5a
Author: Namhyung Kim <namhyung@kernel.org>
Date: Fri Nov 18 11:01:09 2022 -0800

description
===========
It caused some troubles when a lock inside kmalloc is contended
because task local storage would allocate memory using kmalloc.
It'd create a recusion and even crash in my system.

There could be a couple of workarounds but I think the simplest
one is to use a pre-allocated hash map.  We could fix the task
local storage to use the safe BPF allocator, but it takes time
so let's change this until it happens actually.

    Signed-off-by: Namhyung Kim <namhyung@kernel.org>
    Acked-by: Martin KaFai Lau <martin.lau@kernel.org>
    Cc: Adrian Hunter <adrian.hunter@intel.com>
    Cc: Blake Jones <blakejones@google.com>
    Cc: Chris Li <chriscli@google.com>
    Cc: Ian Rogers <irogers@google.com>
    Cc: Ingo Molnar <mingo@kernel.org>
    Cc: Jiri Olsa <jolsa@kernel.org>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Cc: Song Liu <song@kernel.org>
    Cc: bpf@vger.kernel.org
    Link: https://lore.kernel.org/r/20221118190109.1512674-1-namhyung@kernel.org
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>